requirements.txt
fastapi
uvicorn
numpy
faiss-cpu
sentence-transformers
transformers
certifi
